Support cinch, move cinch code to greater lib
